+++ src/main/java/org/example/project/service/subscribetion/SubscriptionService.java
@@ -4,7 +4,7 @@ import org.example.project.entity.subscribetion.Payment;
 import org.example.project.entity.subscribetion.Subscription;
 
 public interface SubscriptionService {
-    Subscription createSubscription(Long userId, Long planId);
+    Subscription createSubscription(Integer userId, Long planId);
     Payment processPayment(Long subscriptionId, Double amount, String paymentMethod);
     void cancelSubscription(Long subscriptionId);
 }
